Move deploy to deployablescontrol
